chore(models): standardize model filenames to PascalCase
