+++ src/models/AutomationRule.js
