+++ src/models/CustomLogo.js
