+++ src/models/index.js
@@ -37,12 +37,6 @@ Object.keys(db).forEach(modelName => {
   }
 });
 
-// Import models explicitly
-const DocumentLayout = require('./documentLayout')(sequelize, Sequelize.DataTypes);
-const ReportTemplate = require('./reporttemplate')(sequelize, Sequelize.DataTypes);
-db.DocumentLayout = DocumentLayout;
-db.ReportTemplate = ReportTemplate;
-
 db.sequelize = sequelize;
 db.Sequelize = Sequelize;
